chore: rename and restructured `WixServerClient` & `WixWebClient` classes
