+++ src/web/index.ts
@@ -1,0 +1,1 @@
+export { WixWebClient } from './wixWebClient';
